Renamed assets and setup project settings
