+++ .github/variables/projectconfig.env
@@ -1,3 +1,3 @@
-PROJECT_NAME=TemplateUnityPackage
+PROJECT_NAME=StateMachineUnity
 UNITY_VERSION=2021.3.11f1
-PACKAGE_PATH=./Packages/com.nickmaltbie.TemplateUnityPackage
+PACKAGE_PATH=./Packages/com.nickmaltbie.StateMachineUnity
