+++ Assets/Editor/AssetGUIDRegenerator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6f16f2d106992a34b8db7e3a2691be98
+guid: 1e70dbee6b970de43a742f42562d8193
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
