+++ Assets/Editor/RenameAssets.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 08ec5285007ef964089d37440fbddca6
+guid: ec00d305d5a8da9469f35a9e1897f051
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
