+++ Assets/Editor/ScriptBatch.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9a90221adc257c145a20030403b3639c
+guid: 5aaca449181214e48938d53de12404f7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
