+++ Assets/Samples/ExampleSample.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ad2e87374c6a1f1449d574c85dde8e3e
+guid: c42f41a974e4b4c409e8456a866cd66f
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
