+++ Assets/Samples/ExampleSample/ExamplePrefab.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 61badf7db01dcf5449d894b1e2fced93
+guid: e81b310fe0b3fc54a87d25521461a012
 PrefabImporter:
   externalObjects: {}
   userData:
