+++ Assets/Samples/ExampleSample/ExampleScene.unity
@@ -308,53 +308,53 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_RootOrder
       value: 2
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalPosition.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalPosition.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalRotation.w
       value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalRotation.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalRotation.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalRotation.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalEulerAnglesHint.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalEulerAnglesHint.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 350952432811950447, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 350952432811950447, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 9179667767173461693, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+    - target: {fileID: 9179667767173461693, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
       propertyPath: m_Name
       value: ExamplePrefab
       objectReference: {fileID: 0}
     m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 61badf7db01dcf5449d894b1e2fced93, type: 3}
+  m_SourcePrefab: {fileID: 100100000, guid: e81b310fe0b3fc54a87d25521461a012, type: 3}
