+++ Assets/Scenes/SampleScene.unity
@@ -330,7 +330,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1654717964}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3af6e1d4a92888b4a80b60119dd406be, type: 3}
+  m_Script: {fileID: 11500000, guid: 8db79e70b9f4ddd4fbac2ad98ecf5fce, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!65 &1654717966
