+++ Documentation/build.sh
@@ -4,7 +4,7 @@ BASEDIR=$(dirname "$0")
 echo "Setting up website and copying files"
 cp $BASEDIR/../README.md $BASEDIR/index.md
 cp $BASEDIR/../LICENSE.txt $BASEDIR/LICENSE.txt
-cp $BASEDIR/../Packages/com.nickmaltbie.TemplateUnityPackage/CHANGELOG.md $BASEDIR/CHANGELOG.md
+cp $BASEDIR/../Packages/com.nickmaltbie.StateMachineUnity/CHANGELOG.md $BASEDIR/CHANGELOG.md
 cp -r $BASEDIR/../Demo $BASEDIR/Demo
 
 # Generate website with docfx
