+++ Documentation/changelog/toc.yml
@@ -1,4 +1,4 @@
-# Changelog is populated from Packages\com.nickmaltbie.TemplateUnityPackage\CHANGELOG.md
+# Changelog is populated from Packages\com.nickmaltbie.StateMachineUnity\CHANGELOG.md
 # as part of build script.
 
 - name: Changelog
