+++ Documentation/docfx.json
@@ -3,7 +3,7 @@
         {
             "src": [
                 {
-                    "src": "../Packages/com.nickmaltbie.TemplateUnityPackage",
+                    "src": "../Packages/com.nickmaltbie.StateMachineUnity",
                     "files": [
                         "**/*.cs"
                     ]
@@ -17,8 +17,8 @@
     ],
     "build": {
         "globalMetadata": {
-            "_appTitle": "Template Unity Package Documentation",
-            "_appFooter": "Template Unity Package Documentation",
+            "_appTitle": "State Machine Unity Documentation",
+            "_appFooter": "State Machine Unity Documentation",
             "_enableSearch": true
         },
         "content": [
@@ -56,7 +56,7 @@
         ],
         "overwrite": [
             {
-                "src": "../Packages/com.nickmaltbie.TemplateUnityPackage",
+                "src": "../Packages/com.nickmaltbie.StateMachineUnity",
                 "files": [
                     "**/*.cs"
                 ]
@@ -73,7 +73,7 @@
             }
         ],
         "sitemap": {
-            "baseUrl": "https://nickmaltbie.com/TemplateUnityPackage/docs",
+            "baseUrl": "https://nickmaltbie.com/StateMachineUnity/docs",
             "changefreq": "weekly",
             "fileOptions": {
                 "latest/api/*": {
