+++ Packages/com.nickmaltbie.StateMachineUnity/CHANGELOG.md.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: fc1f856a37f122e49ad7922398a4541a
-AssemblyDefinitionImporter:
+guid: a1756a4cd8c69574b9783ec4df3e0ee6
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
