+++ Packages/com.nickmaltbie.StateMachineUnity/LICENSE.txt.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 0a260c49d7d6d8046a4b7a23f5460063
-folderAsset: yes
-DefaultImporter:
+guid: c9eac0928af942e4994e22cc11c81db8
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
