+++ Packages/com.nickmaltbie.StateMachineUnity/README.md.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: dc70346e3b16398409055b50b81491c6
-PackageManifestImporter:
+guid: cb0a916dfaa44464cbe79ed9748d443e
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
