+++ Packages/com.nickmaltbie.StateMachineUnity/Scripts.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: faf341d8dc88baa40887e55a646f8429
-AssemblyDefinitionImporter:
+guid: aff28230d3e9fe944a0e8c0e6aab8e56
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
