+++ Packages/com.nickmaltbie.StateMachineUnity/Scripts/SampleScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b64f585020a59f74ba9dc1c718671a8c
+guid: 8db79e70b9f4ddd4fbac2ad98ecf5fce
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
