+++ Packages/com.nickmaltbie.StateMachineUnity/Scripts/nickmaltbie.StateMachineUnity.asmdef.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 218f47f301e83fa47a1e267bc62390f1
-TextScriptImporter:
+guid: ef35993bd7c22e243bb26a52fcfafa36
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
