+++ Packages/com.nickmaltbie.StateMachineUnity/Tests.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 9cc0d73a64630ea4094edf2aa62286e2
-AssemblyDefinitionImporter:
+guid: bd4af458345ed5a4e8f1e224dc5ab41e
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
