+++ Packages/com.nickmaltbie.StateMachineUnity/Tests/EditMode.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: a7a6ad7e1c6a13149b40b7698db96e38
-TextScriptImporter:
+guid: db19a15aea90e9544a2d29ba701cefb2
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
