+++ Packages/com.nickmaltbie.StateMachineUnity/Tests/EditMode/EditModeTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fbb06df6b50ed474ab2e8cb8c7177ebe
+guid: 1eb048cbef0b43a44adae7a133a0afa2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
