+++ Packages/com.nickmaltbie.StateMachineUnity/Tests/EditMode/nickmaltbie.StateMachineUnity.Tests.EditMode.asmdef.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: fbb37ea4ab6d4c94593c8f70ac111607
-TextScriptImporter:
+guid: 58963106e92bd2740a0f6365f8fd016a
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
