+++ Packages/com.nickmaltbie.StateMachineUnity/Tests/PlayMode.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c0e5fe145b1e8af4b8f3d1be0f0bfec7
+guid: ebc9492ae450e74478a41aab5f045380
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
