+++ Packages/com.nickmaltbie.StateMachineUnity/Tests/PlayMode/PlayModeTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3af6e1d4a92888b4a80b60119dd406be
+guid: d9cc1ad396f562e4fad600513093b76e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
