+++ Packages/com.nickmaltbie.StateMachineUnity/Tests/PlayMode/nickmaltbie.StateMachineUnity.Tests.PlayMode.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: f9165928ac427214d886d8162892680f
-folderAsset: yes
-DefaultImporter:
+guid: 1e009cb58b2871442a32dc14910d0f62
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
