+++ Packages/com.nickmaltbie.StateMachineUnity/package.json.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: f8fc89548295c2546b0685bed0925b52
-folderAsset: yes
-DefaultImporter:
+guid: 30c6d3d8395e3364ebab81c908975fe7
+PackageManifestImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
